function rename to getResolverOptions
